Change instructions to use `git config`
Due to each installation requiring a different `user.email`, it would be better
to set-up the config file through `git config`.  This would reduce the chances
of accidentally copying the file from the repo without changing the name.

In the future, when the install script sets `/etc/hostname`, it can also set
an associated email somewhere in the system.  The git install script can then
pipe that value into the config file automatically.